requirements.txt
langchain==0.3.12
langchain-community==0.3.12
langchain-google-genai==2.0.7
python-dotenv==1.0.1
pillow==11.0.0
tqdm>=4.65.0
transformers
torch
faiss-cpu